Allow to specify the JTS GeometryFactory used to build geometries obtained from RevFeature.
When using FormatCommon v2.1, RevFeature values are lazily
parsed out of a member byte array.
By the other hand, the GeoTools Renderer is more efficient
if the data source respects the GeometryFactory is gives as
a hint.
The combination of the two makes it possible for the RevFeature
implementation to use the requested GeometryFactory without
affecting the RevFeature's immutable state.

Signed-off-by: Gabriel Roldan <groldan@boundlessgeo.com>